Update where.md to enhance the usability of the example
The example on the page is useful; however, it is rigid in only allowing a 3 day look back. Adjusting the pattern to include a number value and a macro that performs a regex search to set the dateadd logic to whatever value is configured in the test. This allows for reusbaility of the macro for different lookback days.